Add files via upload
The dataset consists of ~45,000 rows for ~20 baking products & customers datapoints. Dataset consists of direct marketing campaigns by a Portuguese banking institution, using phone calls, which are aimed to sell subscriptions to a bank term deposit.

The Application is a recommendation system with a backend & frontend to suggest if the customer will buy the product or not.